requirements.txt
httpx
gspread
google-auth
python-dotenv
